+++ lib/screens/daily_rewards.dart
@@ -1,6 +1,5 @@
 ﻿import 'dart:math';
 import 'package:confetti/confetti.dart';
-import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -86,8 +85,8 @@ class DailyRewardDialog {
 
     final (xpGained, baseXp, streak, multiplier, levelBefore) = result;
 
-    FirebaseAnalytics.instance.logEvent(
-      name: 'daily_reward_claimed',
+    logAnalyticsEvent(
+      'daily_reward_claimed',
       parameters: {
         'xp_gained': xpGained,
         'streak': streak,